Initial commit for usecase#1 Priority aware admission control.
Create a custom resource:  OS::Nova::AdmissionControl
In every stack template corresponding to the customer's service request,
include a resource of this type 'OS::Nova::AdmissionControl'.

This results in an credit mechanism for the service.
The resource maintains a per resource class {gold/silver/bronze} usage
and threshold.

Stack creation is permitted only if the resource's class is within its
threshold.